requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
dnspython==2.7.0
easydict==1.13
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
h11==0.16.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.1
onnxruntime-silicon==1.16.3
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.30.2
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
